# backend/swagger_server/swagger/swagger.yaml
openapi: 3.0.3
info:
  title: Mosaic API
  description: "Mosaic will be a social media platform designed for individuals to\
    \ engage in discussions and share content in a community-driven environment. Similar\
    \ to Reddit, Mosaic will allow users to create and join communities that are focused\
    \ on specific topics of interest. As the name suggests, our aim is to bring together\
    \ a diverse community made up of different pieces that come together to create\
    \ a beautiful and meaningful whole."
  termsOfService: http://link_to_mosaic/terms/
  contact:
    email: apiteam@mosaic.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: "1.0"
...
servers:
- url: https://link_to_mosaic/api/v3
...
paths:
...
  /u/create:
    post:
      tags:
      - user
      summary: Create user
      description: This can only be done by the logged in user.
      operationId: create_user
      requestBody:
        description: Created user object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        default:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
      x-openapi-router-controller: swagger_server.controllers.user_controller
  /u/login:
    post:
      tags:
      - user
      summary: Logs user into the system
      description: ""
      operationId: login_user
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/u_login_body'
        required: true
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: string
                x-content-type: application/json
        "400":
          description: Invalid username/password supplied
      x-openapi-router-controller: swagger_server.controllers.user_controller
  /u/login/forgotPassword:
    get:
      tags:
      - user
      summary: Sends recovery email
      description: ""
      operationId: forgot_password
      parameters:
      - name: username
        in: query
        description: The user name for login
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: email
        in: query
        description: The email for login
        required: false
        style: form
        explode: true
        schema:
          type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: string
                x-content-type: application/json
        "400":
          description: Invalid username/email supplied
      x-openapi-router-controller: swagger_server.controllers.user_controller
  /u/{username}:
    get:
      tags:
      - user
      summary: Get user by user name
      description: ""
      operationId: get_user_by_name
      parameters:
      - name: username
        in: path
        description: The name of the user. Use reaper_113 for testing.
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "400":
          description: Invalid username supplied
        "404":
          description: User not found
      x-openapi-router-controller: swagger_server.controllers.user_controller
    put:
      tags:
      - user
      summary: Update user
      description: This can only be done by the logged in user.
      operationId: update_user
      parameters:
      - name: username
        in: path
        description: The name of the user. Use reaper_113 for testing.
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        description: Update an existent user in the store
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        default:
          description: successful operation
      security:
      - user_auth:
        - login:user
      x-openapi-router-controller: swagger_server.controllers.user_controller
    delete:
      tags:
      - user
      summary: Delete user
      description: This can only be done by the logged in user.
      operationId: delete_user
      parameters:
      - name: username
        in: path
        description: The name of the user. Use reaper_113 for testing.
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "400":
          description: Invalid username supplied
        "404":
          description: User not found
      security:
      - user_auth:
        - login:user
      x-openapi-router-controller: swagger_server.controllers.user_controller
  /u/{username}/notif:
    get:
      tags:
      - user
      summary: Get user notifications
      description: ""
      operationId: get_user_notifications
      parameters:
      - name: username
        in: path
        description: The name of the user. Use reaper_113 for testing.
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notifications'
        "400":
          description: Invalid username supplied
        "404":
          description: User not found
      security:
      - user_auth:
        - login:user
      x-openapi-router-controller: swagger_server.controllers.user_controller
  /u/{username}/pref:
    get:
      tags:
      - user
      summary: Get user preferences
      description: ""
      operationId: get_user_preferences
      parameters:
      - name: username
        in: path
        description: The name of the user. Use reaper_113 for testing.
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Preferences'
        "400":
          description: Invalid username supplied
        "404":
          description: User not found
      security:
      - user_auth:
        - login:user
      x-openapi-router-controller: swagger_server.controllers.user_controller
    put:
      tags:
      - user
      summary: Update user preferences
      description: This can only be done by the logged in user.
      operationId: update_user_preferences
      parameters:
      - name: username
        in: path
        description: The name of the user. Use reaper_113 for testing.
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        description: Update prefrences of a user
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Preferences'
      responses:
        default:
          description: successful operation
      security:
      - user_auth:
        - login:user
      x-openapi-router-controller: swagger_server.controllers.user_controller
...
components:
  schemas:
...
    User:
      required:
      - emailAddress
      - password
      - username
      type: object
      properties:
        UserID:
          type: integer
          format: int64
          example: 1012312
        username:
          type: string
          example: reaper_113
        emailAddress:
          type: string
          example: reaper@email.com
        password:
          type: string
          example: zxcvbnm
        timeJoined:
          type: string
          format: date-time
        lastLogin:
          type: string
          format: date-time
        userStatus:
          type: integer
          description: User Status
          format: int32
          example: 1
        profilePicture:
          type: string
        karma:
          type: integer
          format: int32
        userPreferences:
          type: integer
          format: int64
      example:
        lastLogin: 2000-01-23T04:56:07.000+00:00
        profilePicture: profilePicture
        karma: 0
        emailAddress: reaper@email.com
        password: zxcvbnm
        userStatus: 1
        timeJoined: 2000-01-23T04:56:07.000+00:00
        UserID: 1012312
        username: reaper_113
      xml:
        name: user
...
    Notifications:
      type: object
      properties:
        notificationID:
          type: integer
          format: int64
        content:
          type: string
        UserID:
          type: integer
          format: int64
        postID:
          type: integer
          format: int64
        communityID:
          type: integer
          format: int64
        commentID:
          type: integer
          format: int64
        messageID:
          type: integer
          format: int64
      example:
        postID: 1
        commentID: 5
        notificationID: 0
        communityID: 5
        messageID: 2
        userID: 6
        content: content
...
    Preferences:
      type: object
      properties:
        preferenceID:
          type: integer
          format: int64
        sortHome:
          type: string
          enum:
          - Best
          - New
          - Controversial
          - Old
        time:
          type: string
          format: date-time
        sortTrending:
          type: string
          enum:
          - Best
          - New
          - Controversial
          - Old
        sortCommunities:
          type: string
          enum:
          - Best
          - New
          - Controversial
          - Old
        language:
          type: string
          default: English
          enum:
          - English
        blurNSFW:
          type: boolean
          default: true
        autoplay:
          type: boolean
          default: true
        darkMode:
          type: boolean
          default: false
      example:
        sortHome: Best
        blurNSFW: true
        language: English
        preferenceID: 0
        darkMode: false
        sortCommunities: Best
        time: 2000-01-23T04:56:07.000+00:00
        sortTrending: Best
        autoplay: true
      xml:
        name: preferences
...
    u_login_body:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
...
  securitySchemes:
    user_auth:
      type: oauth2
      flows:
        password:
          tokenUrl: https://link_to_mosaic.com/oauth/authorize
          scopes:
            login:user: user logged in
      x-tokenInfoFunc: swagger_server.controllers.authorization_controller.check_user_auth
      x-scopeValidateFunc: swagger_server.controllers.authorization_controller.validate_scope_user_auth
    api_key:
      type: apiKey
      name: api_key
      in: header
      x-apikeyInfoFunc: swagger_server.controllers.authorization_controller.check_api_key